Add support to Mark as Done
Rename the class static fields to align with new class: "Task".
Create new static method to handle mark tasks as done or undone.
Use regression to match patterns: "mark" or "unmark" followed by a number.
Handle negative number input or number that is bigger than the size of
the list in the new class method.